--- expense/.env.example
@@ -1,2 +1,0 @@
-DATABASE_URL=
-JWT_SECRET=
