--- expense/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "expense",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "node src/server",
-    "start:watch": "node --watch src/server",
-    "swagger": "node src/swagger",
-    "swagger:watch": "node --watch src/swagger"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "module",
-  "description": "",
-  "dependencies": {
-    "@prisma/client": "^6.8.2",
-    "dotenv": "^16.5.0",
-    "expense": "file:",
-    "express": "^5.1.0",
-    "http-status": "^2.1.0",
-    "jsonwebtoken": "^9.0.2",
-    "swagger-autogen": "^2.23.7",
-    "yup": "^1.6.1"
-  },
-  "devDependencies": {
-    "prisma": "^6.8.2"
-  }
-}
