--- expense/prisma/schema.prisma
@@ -1,19 +1,0 @@
-datasource db {
-  provider = "mysql"
-  url      = env("DATABASE_URL")
-}
-
-generator client {
-  provider      = "prisma-client-js"
-  binaryTargets = ["native", "windows", "debian-openssl-3.0.x"]
-}
-
-model Expense {
-  expenseId       Int      @id @default(autoincrement())
-  userId          Int
-  value           Float
-  categoryId      Int
-  paymentMethodId Int
-  expenseDate     DateTime
-  description     String?
-}
