+++ paymentMethod/.env.example
@@ -1,2 +1,3 @@
 DATABASE_URL=
 JWT_SECRET=
+JWT_EXPIRE=
