+++ transaction/src/application/TransactionService.js
@@ -1,0 +1,9 @@
+import TransactionRepository from '../infrastructure/database/TransactionRepository.js';
+
+export default {
+  findAll: TransactionRepository.findAll,
+  findById: TransactionRepository.findById,
+  create: TransactionRepository.create,
+  update: TransactionRepository.update,
+  remove: TransactionRepository.remove,
+};
