+++ transaction/src/swagger.js
@@ -1,0 +1,40 @@
+import swaggerAutogen from 'swagger-autogen';
+
+const doc = {
+  info: {
+    version: '1.0.0',
+    title: 'Tecnologias Emergentes',
+    description: 'Documentação da API criada em sala',
+  },
+  servers: [
+    {
+      url: 'http://localhost:4041/',
+    },
+  ],
+  components: {
+    schemas: {
+      InternalServerError: {
+        code: '',
+        message: '',
+      },
+      User: {
+        name: '',
+        email: '',
+        password: '',
+      },
+    },
+    securitySchemes: {
+      bearerAuth: {
+        type: 'http',
+        scheme: 'bearer',
+      },
+    },
+  },
+};
+
+const outputFile = './config/swagger.json';
+const endpointsFiles = ['./server.js'];
+
+swaggerAutogen({ openapi: '3.0.0' })(outputFile, endpointsFiles, doc).then(async () => {
+  await import('./server.js');
+});
